+++ packages/flow-pages/composables/avue.ts
@@ -1,9 +1,11 @@
 import type { AvueCrudOption } from '@smallwei/avue'
 
-import { reactive } from 'vue'
+import { reactive, watch } from 'vue'
 import { useConfigProvider } from '@yusui/flow-pages'
 
 export function defineTableOption<T>(option: AvueCrudOption<T>) {
   const { tableOption } = useConfigProvider()
-  return reactive({ ...tableOption, ...option } as AvueCrudOption<T>)
+  const reactiveOption = reactive<AvueCrudOption<T>>({ ...tableOption, ...option })
+  watch(tableOption!, () => Object.assign(reactiveOption, tableOption))
+  return reactiveOption
 }