+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## [2.0.9](https://github.com/SoulLyoko/yusui-tools/compare/v2.0.8...v2.0.9) (2024-02-21)
+
+
+### Bug Fixes
+
+* **IconSelect:** 处理collections为匹配方式时的问题 ([e4b6da9](https://github.com/SoulLyoko/yusui-tools/commit/e4b6da935d9dbcbde20086344097945d3e4b3450))
+* **useDict:** 内部实现toValue ([6040b4a](https://github.com/SoulLyoko/yusui-tools/commit/6040b4acec5dcf6b802837c063b18d994114b0c8))
+
+
+
+## [2.0.8](https://github.com/SoulLyoko/yusui-tools/compare/v2.0.7...v2.0.8) (2024-02-21)
+
+
+### Bug Fixes
+
+* **flow-form:** 非待办时禁用表单 ([58c1e17](https://github.com/SoulLyoko/yusui-tools/commit/58c1e1778ab0078053628ea7fa60bb4b9a142bc1))
+* **flow-form:** 修复在onBeforeSubmit修改表单值导致审批窗口重新加载的问题 ([8088646](https://github.com/SoulLyoko/yusui-tools/commit/8088646c11222f20d8bd30ec78906c92c9f685e8))
+* **flow-form:** 修复afterGetDetail不触发的问题 ([5055e72](https://github.com/SoulLyoko/yusui-tools/commit/5055e727447dc0dba007db0575728ed0e596e66a))
+* **IconSelect:** tabs遵循collections的排序 ([71a294f](https://github.com/SoulLyoko/yusui-tools/commit/71a294f1e568657daba57b499b30b261632232f3))
+* **useDict:** 修复dicData丢失响应的问题 ([2391056](https://github.com/SoulLyoko/yusui-tools/commit/2391056b23dddf2fb01f86db0d373af43db022b9))
+
+
+### Features
+
+* **flow-form:** 审批树已选择项过多时以弹出的形式展示剩余项 ([050aee0](https://github.com/SoulLyoko/yusui-tools/commit/050aee02f0c47efe5a571aac9abb594e8cc23bf1))
+* **flow-pages:** 从useProvideState中导出流程参数和按钮处理 ([e67da1d](https://github.com/SoulLyoko/yusui-tools/commit/e67da1d7a788a17ef175cb1e8383835874cd9801))
+
+
+
+## [2.0.7](https://github.com/SoulLyoko/yusui-tools/compare/v2.0.6...v2.0.7) (2024-01-29)
+
+
+### Bug Fixes
+
+* **flow-form:** 优化移动端滑动 ([fb318af](https://github.com/SoulLyoko/yusui-tools/commit/fb318af059df8b695472daf9438fd39f9982f415))
+* **flow-pages:** 修复表格配置合并 ([c85b105](https://github.com/SoulLyoko/yusui-tools/commit/c85b1059d7fe3bae0fd2b618b5c6d76397554b4a))
+* **form-design:** 组件不可见时调低透明度 ([29d7ff4](https://github.com/SoulLyoko/yusui-tools/commit/29d7ff4d69fd000a2f774ba2482f9a76b28cbc51))
+
+
+
 ## [2.0.6](https://github.com/SoulLyoko/yusui-tools/compare/v2.0.4...v2.0.6) (2024-01-09)